requirements1.txt
#python==3.10.11
langchain_groq==0.2.1
pypdf==5.1.0
langchain_community==0.3.5
langchain-huggingface==0.1.2
langchain_chroma==0.1.4
langchain-text-splitters==0.3.2
streamlit==1.40.0
python-dotenv==1.0.1
fastapi==0.115.4
uvicorn==0.32.0